dist/Makefile: Use 'BINDIR' for APPDEPS
Files should not be generated in the source but in the build directory.
So show more this as an example.

Since now `Makefile.include` defines ".DEFAULT_GOAL := all" there is no
limitation anymore on 'otherwise you modify the standard target'

It is defined after Makefile.include to use the BINDIR variable.