boards/common/nrf52: add openocd support for 'nordic_softdevice_ble'
Enable the handling of flashing `softdevice.hex` when flashing the firmware
for openocd.

However, for flashing, only the `hexfile` and `binfile` can currently be used.

The `elffile` is generated with local pages aligned to `0x10000` which makes
the program starting at `0x1f000` be flashed from `0x10000` with padding bytes
even if the `.text` section is indeed at `0x1f000`:

    readelf --sections bin/nrf52dk/gnrc_networking.elf

      ...
      [ 1] .text         PROGBITS    0001f000 00f000 00f698 00  AX  0   0 16
      ...

    readelf --segments  bin/nrf52dk/gnrc_networking.elf

      ...
      LOAD           0x000000 0x00010000 0x00010000 0x1e6a0 0x1e6a0 R E 0x10000
      ...

The padding bytes would go through `verify_image` in `openocd` so be expected
to not be overwritten but are by `softdevice.hex`

Using --nmagic at link time removes the local page alignement but would
need dedicated testing.