tests/libfixmath: increase test timeout
It is a test that outputs a lot of text which can be slow to handle.

Tests were taking 50 seconds when I re-tested on boards that failed during
release testing because of timeout on my computer.
With an a bit loaded computer it may fail because of not handling output fast
enough so increase the default timeout.