TODO IMPROVE DOCUMENTATION Makefile.base: SUBMODULES as list of all submodules
Add a new handling for `SUBMODULES` to be declared as a list of the
submodules that should be detected. The other source files will by
default be automatically used as considered part of the main module.

It allows having pseudomodules implemented in a module that also has
submodules. It would have before led to missing source files or enabling
'SUBMODULES_NOFORCE' which would remove checking for correct module
existence.

It is also allowed to use 'SUBMODULES = $(BASE_MODULE)_%' as wildcard to
mimic the previous behavior and consider all files as possible submodule.